Tidy up and extract scale bar component
